requirements.txt
python-telegram-bot==21.11.1
fastapi>=0.111
uvicorn[standard]>=0.30
python-dotenv>=1.0

